+++ example/android/app/src/main/kotlin/com/mukarrom/example/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.mukarrom.example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity()