+++ CHANGELOG.md
@@ -1,3 +1,12 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Color Selector Widget Added.
+
+## 1.0.0
+
+* Size Selector Widget Added.
+
+## 1.1.0
+
+* Package name changed.
+* Fixed some minor bugs.